Create Budgeting pages (#15)
This creates static pages for the budgeting pages.

I thought it was better to do it this way just so it
would be easier to make the content dynamic.

These pages are mobile-friendly as well. I figured
since it would take a while before I can create a
mobile app, this would be a good way to make
the app usable on mobile.